Move to separate market_models file
